Improve Iceberg commit to Glue
Glue uses an optional versionId during updates to make
sure there are no conflicts. If the given version has
already been updated, the Iceberg commit will retry.
