+++ plugin/trino-iceberg/src/test/java/io/trino/plugin/iceberg/BaseIcebergConnectorSmokeTest.java
@@ -20,9 +20,20 @@ import io.trino.testing.sql.TestTable;
 import org.apache.iceberg.FileFormat;
 import org.testng.annotations.Test;
 
+import java.util.List;
+import java.util.concurrent.CyclicBarrier;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Future;
+import java.util.stream.IntStream;
+
+import static com.google.common.collect.ImmutableList.toImmutableList;
+import static io.airlift.concurrent.MoreFutures.getFutureValue;
 import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
+import static java.util.concurrent.Executors.newFixedThreadPool;
+import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.testng.Assert.assertTrue;
 
 public abstract class BaseIcebergConnectorSmokeTest
         extends BaseConnectorSmokeTest
@@ -83,6 +94,48 @@ public abstract class BaseIcebergConnectorSmokeTest
 
             // Check whether the "$path" hidden column is correctly evaluated in the filter expression
             assertQuery(format("SELECT a FROM %s WHERE \"$path\" = '%s'", table.getName(), filePath), "VALUES 1");
+        }
+    }
+
+    // Repeat test with invocationCount for better test coverage, since the tested aspect is inherently non-deterministic.
+    @Test(timeOut = 60_000, invocationCount = 4)
+    public void testDeleteRowsConcurrently()
+            throws Exception
+    {
+        int threads = 4;
+        CyclicBarrier barrier = new CyclicBarrier(threads);
+        ExecutorService executor = newFixedThreadPool(threads);
+        try (TestTable table = new TestTable(
+                getQueryRunner()::execute,
+                "test_concurrent_delete",
+                "(col0 INTEGER, col1 INTEGER, col2 INTEGER, col3 INTEGER)")) {
+            String tableName = table.getName();
+            assertUpdate("INSERT INTO " + tableName + " VALUES (0, 0, 0, 0)", 1);
+            assertUpdate("INSERT INTO " + tableName + " VALUES (1, 0, 0, 0)", 1);
+            assertUpdate("INSERT INTO " + tableName + " VALUES (0, 1, 0, 0)", 1);
+            assertUpdate("INSERT INTO " + tableName + " VALUES (0, 0, 1, 0)", 1);
+            assertUpdate("INSERT INTO " + tableName + " VALUES (0, 0, 0, 1)", 1);
+
+            List<Future<Boolean>> futures = IntStream.range(0, threads)
+                    .mapToObj(threadNumber -> executor.submit(() -> {
+                        barrier.await(10, SECONDS);
+                        try {
+                            String columnName = "col" + threadNumber;
+                            getQueryRunner().execute(format("DELETE FROM %s WHERE %s = 1", tableName, columnName));
+                            return true;
+                        }
+                        catch (Exception e) {
+                            return false;
+                        }
+                    }))
+                    .collect(toImmutableList());
+
+            futures.forEach(future -> assertTrue(getFutureValue(future)));
+            assertThat(query("SELECT max(col0), max(col1), max(col2), max(col3) FROM " + tableName)).matches("VALUES (0, 0, 0, 0)");
+        }
+        finally {
+            executor.shutdownNow();
+            executor.awaitTermination(10, SECONDS);
         }
     }
 }
